+++ src/UrlGenerator/UrlGeneratorFactory.php
@@ -11,8 +11,12 @@ class UrlGeneratorFactory
 {
     public static function createForMedia(Media $media, string $conversionName = ''): UrlGenerator
     {
+        $diskDriverName = ($conversionName === '')
+            ? $media->getDiskDriverName()
+            : $media->getConversionsDiskDriverName();
+
         $urlGeneratorClass = config('medialibrary.url_generator')
-            ?: 'Spatie\MediaLibrary\UrlGenerator\\'.ucfirst($media->getDiskDriverName()).'UrlGenerator';
+            ?: 'Spatie\MediaLibrary\UrlGenerator\\'.ucfirst($diskDriverName).'UrlGenerator';
 
         static::guardAgainstInvalidUrlGenerator($urlGeneratorClass);