+++ tests/Support/TestModels/TestModelWithConversionsOnOtherDisk.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Spatie\MediaLibrary\Tests\Support\TestModels;
+
+use Spatie\MediaLibrary\Models\Media;
+
+class TestModelWithConversionsOnOtherDisk extends TestModel
+{
+    /**
+     * Register the conversions that should be performed.
+     *
+     * @return array
+     */
+    public function registerMediaConversions(Media $media = null)
+    {
+        $this->addMediaConversion('thumb')
+            ->withResponsiveImages()
+            ->width(50)
+            ->nonQueued();
+    }
+
+    /**
+     * Register the media collections and conversions.
+     *
+     * @return void
+     */
+    public function registerMediaCollections()
+    {
+        $this
+            ->addMediaCollection('thumb')
+            ->storeConversionsOnDisk('secondMediaDisk');
+    }
+}